Fix circular dependency in copy-manager
